Break Client views files a little more
