+++ Client/Client/SomethingView.swift
@@ -1,0 +1,11 @@
+import SwiftUI
+
+struct SomethingView: View {
+    var body: some View {
+        VStack {
+            Text("Something View")
+                .accessibilityIdentifier("Something View")
+                .accessibilityLabel("SomethingViewAccessbilityLabel")
+        }
+    }
+}
