+++ Client/Client/StringView.swift
@@ -1,0 +1,11 @@
+import SwiftUI
+
+struct StringView: View {
+    var value: String
+
+    var body: some View {
+        VStack {
+            Text("Value: \(value)")
+        }
+    }
+}
